@@ -5,19 +5,13 @@ using System.Data.SqlClient;
 
 namespace Napper.Database
 {
-    public class MySqlDataAccessor : IDataAccessor
+    public class MySqlDataAccessor(string connectionString) : IDataAccessor
     {
         MySqlConnection? _connection = null;
         MySqlCommand? _command = null;
-        string _connectionString = "";
+        private readonly string _connectionString = connectionString;
 
         public ConnectionState State => (_connection == null) ? ConnectionState.Closed : _connection.State;
-
-
-        public MySqlDataAccessor(string connectionString)
-        {
-            _connectionString = connectionString;
-        }
 
         public bool TryOpenConnection(out string errorMessage)
         {
@@ -129,7 +123,7 @@ namespace Napper.Database
                 {
                     foreach (var parameter in parameters)
                     {
-                        if (fields.ContainsKey(parameter.Key))
+                        if (fields.TryGetValue(parameter.Key, out Type? value))
                         {
                             if (whereSql == "")
                             {
@@ -139,7 +133,7 @@ namespace Napper.Database
                             {
                                 whereSql += " AND " + parameter.Key + " = @" + parameter.Key;
                             }
-                            _command.Parameters.Add(new MySqlParameter("@" + parameter.Key, Convert.ChangeType(parameter.Value, fields[parameter.Key])));
+                            _command.Parameters.Add(new MySqlParameter("@" + parameter.Key, Convert.ChangeType(parameter.Value, value)));
                         }
                         else
                         {
@@ -176,7 +170,7 @@ namespace Napper.Database
                 throw new NullReferenceException(nameof(TrySelectExecute));
 
             errorMessage = string.Empty;
-            result = new List<Dictionary<string, object>>();
+            result = [];
 
             try
             {
